% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/405_optim_generalized_Bayes_Y_x_with_GGM.R
\name{BTR_Y_x_with_GGM}
\alias{BTR_Y_x_with_GGM}
\title{Generalized-Dimension Bayesian Tensor Response Regression with Gaussian Graphical Model}
\usage{
BTR_Y_x_with_GGM(input, n.iter, n.burn, ranks, hyperparameters = NULL,
  save_after = NULL, save_llik = TRUE, results_list = NULL)
}
\arguments{
\item{input}{A list with at least the elements *Y* and *x*. *Y* should be a list of *G* arrays of dimension $D + 2$ in which the last two dimension margins represent the time and the subject, respectively.}

\item{n.iter}{The number of iterations that the MCMC should run}

\item{n.burn}{The number of iterations at the beginning of the sampler that should be discarded as a "burn-in"}

\item{ranks}{The number of ranks in the model. See Spencer et al. [2019] for details.}

\item{hyperparameters}{A data frame with one observation and eight variables: *a.tau*, *b.tau*, *a.lambda*, *b.lambda*, *a.zeta*, *b.zeta*, *a.sig*, and *b.sig*. Each should specify values to use for the hyperparameters in the model. If `NULL`, then the default values of a.tau = D - 1, b.tau = ranks^((1 / D) - 1), a.lambda = 3, b.lambda = 3^(1/(2*D)), a.zeta = 1, b.zeta = 0.01, a.sig = 1, b.sig = -log(0.95) are used.}
}
\value{
A list with 12 elements, *B*, *W*, *lambda*, *Phi*, *tau*, *d*, *zeta*, *Sig*, *sig2y*, *alpha*, *accept*, and *llik*. *B* is a list of length *n.iter* - *n.burn*, each with $G$ sublists. This will be improved in the future. *llik* is a vector of the log-likelihood values at each MCMC iteration.
}
\description{
Generalized-Dimension Bayesian Tensor Response Regression with Gaussian Graphical Model
}
\examples{
BTRR_model_results <- BTR_Y_x_with_GGM(simulated_fmri_data,10,0,1)

}
