% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/403_vectorized_model.R
\name{vec_Y_x_with_GGM}
\alias{vec_Y_x_with_GGM}
\title{Vectorized Tensor Regression with Gaussian Graphical Model}
\usage{
vec_Y_x_with_GGM(input, n.iter, n.burn, hyperparameters = NULL)
}
\arguments{
\item{input}{A list with at least the elements *Y* and *x*. *Y* should be a list of *G* arrays of dimension $D + 2$ in which the last two dimension margins represent the time and the subject, respectively.}

\item{n.iter}{The number of iterations that the MCMC should run}

\item{n.burn}{The number of iterations at the beginning of the sampler that should be discarded as a "burn-in"}

\item{hyperparameters}{A data frame with one observation and eight variables: *a.tau*, *b.tau*, *a.lambda*, *b.lambda*, *a.zeta*, *b.zeta*, *a.sig*, and *b.sig*. Each should specify values to use for the hyperparameters in the model. If `NULL`, then the default values used are a.tau = D - 1 b.tau = 1^((1 / D) - 1), a.lambda = 3, b.lambda = 3^(1/(2*D)), a.zeta = 1, b.zeta = 0.01, a.sig = 1, b.sig = -log(0.95).}
}
\value{
A list with 9 elements, *B*, *W*, *lambda*, *tau*, *d*, *zeta*, *Sig*, *sig2y*, and *llik*. *B* is a list of length *n.iter* - *n.burn*, each with $G$ sublists. This will be improved in the future. *llik* is a vector of the log-likelihood values at each MCMC iteration.
}
\description{
Vectorized Tensor Regression with Gaussian Graphical Model
}
