% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_Y_x_with_GGM_data.R
\name{Y_x_with_GGM_data}
\alias{Y_x_with_GGM_data}
\title{Simulate data for the Bayesian Tensor Response Regression with Gaussian Graphical Modeling}
\usage{
Y_x_with_GGM_data(subjects = 50, regions = 5, max_time = 100,
  avg_margin_size = c(10, 10, 10), SNR = 1, CNR = 1,
  conn_regions = 1, conn_level = 0.95)
}
\arguments{
\item{subjects}{A scalar. The number of subjects in the sample.}

\item{regions}{A scalar. The number of distinct tensor responses for each subject at each time step.}

\item{max_time}{A scalar. The number of time steps to be simulated from.}

\item{avg_margin_size}{A vector of length *D* that gives the average margin lengths for the tensor reponse *regions* for a given time and subject.}

\item{SNR}{A scalar. The signal-to-noise ratio as defined by Welvaert and Rosseel, 2013}

\item{CNR}{A scalar. The contrast-to-noise ratio as defined by Welvaert and Rosseel, 2013}

\item{conn_regions}{A scalar. The number of pairs of regions that have significant connectivity.}

\item{conn_level}{A scalar. The strength of the correlation between the connected regions, between 0 and 1}
}
\value{
A list with five elements, the response *Y*, the covariate *x*, and the true values for $d_{i,g}$, $\mathbf{B}_g$ and $\boldsymbol{\Sigma}$.
}
\description{
Simulate data for the Bayesian Tensor Response Regression with Gaussian Graphical Modeling
}
\examples{
simulated_data <- Y_x_with_GGM_data()
}
